backend/user-service/requirements.txt
fastapi
pymongo
python-dotenv
uvicorn
requests
python-multipart
passlib
email-validator
